requirements.txt
about-time==4.2.1
alive-progress==3.1.5
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.0
grapheme==0.6.0
ImageHash==4.3.1
imageio==2.33.1
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.5.0
scikit-image==0.22.0
scipy==1.11.4
six==1.16.0
tifffile==2023.12.9
